rename api to sanoty
